+++ src/test/scala/au/com/agl/arc/util/ConfigUtilsSuite.scala
@@ -335,9 +335,92 @@ hdfs://test/{ab,c{de, fg}
 
     pipeline match {
       case Left(stageError) => {
-        assert(stageError == StageError("file extract",3,List(ConfigError("delimiter", Some(12), "Invalid value. Valid values are ['Comma','Pipe','DefaultHive']."))) :: Nil)
+        assert(stageError == StageError("file extract",3,List(ConfigError("delimiter", Some(12), "Invalid value. Valid values are ['Comma','Pipe','DefaultHive','Custom']."))) :: Nil)
       }
       case Right(_) => assert(false)
     }
   }  
+
+  test("Test read custom delimiter") { 
+    implicit val spark = session
+    implicit val logger = LoggerFactory.getLogger(spark.sparkContext.applicationId)
+    implicit val arcContext = ARCContext(jobId=None, jobName=None, environment="test", environmentId=None, configUri=None, isStreaming=false, ignoreEnvironments=false)
+
+    val conf = """{
+      "stages": [       
+        {
+          "type": "DelimitedExtract",
+          "name": "file extract",
+          "environments": [
+            "production",
+            "test"
+          ],
+          "inputView": "input",
+          "outputView": "output",
+          "delimiter": "Custom"
+        }
+      ]
+    }"""
+
+    val base = ConfigFactory.load()
+    val etlConf = ConfigFactory.parseString(conf, ConfigParseOptions.defaults().setSyntax(ConfigSyntax.CONF))
+    val config = etlConf.withFallback(base)
+    var argsMap = collection.mutable.Map[String, String]()
+    val pipeline = ConfigUtils.readPipeline(config.resolve(), new URI(""), argsMap, arcContext)    
+
+    pipeline match {
+      case Left(stageError) => {
+        assert(stageError == StageError("file extract",3,List(ConfigError("customDelimiter", None, "Missing required attribute 'customDelimiter'."))) :: Nil)
+      }
+      case Right(_) => assert(false)
+    }
+  }    
+
+  test("Test read custom delimiter success") { 
+    implicit val spark = session
+    implicit val logger = LoggerFactory.getLogger(spark.sparkContext.applicationId)
+    implicit val arcContext = ARCContext(jobId=None, jobName=None, environment="test", environmentId=None, configUri=None, isStreaming=false, ignoreEnvironments=false)
+
+    val conf = """{
+      "stages": [       
+        {
+          "type": "DelimitedExtract",
+          "name": "file extract",
+          "environments": [
+            "production",
+            "test"
+          ],
+          "inputView": "input",
+          "outputView": "output",
+          "delimiter": "Custom",
+          "customDelimiter": "%"
+        }
+      ]
+    }"""
+
+    val base = ConfigFactory.load()
+    val etlConf = ConfigFactory.parseString(conf, ConfigParseOptions.defaults().setSyntax(ConfigSyntax.CONF))
+    val config = etlConf.withFallback(base)
+    var argsMap = collection.mutable.Map[String, String]()
+    val pipeline = ConfigUtils.readPipeline(config.resolve(), new URI(""), argsMap, arcContext)    
+
+
+    val expected = ETLPipeline(      
+      DelimitedExtract(
+        name="file extract",
+        cols=Right(Nil),
+        outputView="output",
+        input=Left("input"),
+        settings=new Delimited(header=false, sep=Delimiter.Custom, inferSchema=false, customDelimiter="%"),
+        authentication=None,
+        params=Map.empty,
+        persist=false,
+        numPartitions=None,
+        partitionBy=Nil,
+        contiguousIndex=true
+      ) :: Nil)
+
+
+    assert(pipeline === Right(expected))
+  }    
 }